model/requirements.txt
flask>=3.0.0
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.27.0
sentencepiece
protobuf>=4.21.0
python-dotenv
requests
urllib3
azure-cognitiveservices-speech
flask-cors
